msg: 2025-07-05 codes
