+++ alembic/versions/33cbde1cf432_added_remarks_column_for_both_.py
@@ -1,0 +1,34 @@
+"""added remarks column for both endorsement tables
+
+Revision ID: 33cbde1cf432
+Revises: e7c622ff2ce8
+Create Date: 2025-07-05 14:18:16.746170
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '33cbde1cf432'
+down_revision: Union[str, Sequence[str], None] = 'e7c622ff2ce8'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('tbl_endorsement_t1', sa.Column('t_remarks', sa.String(length=100), nullable=True))
+    op.add_column('tbl_endorsement_t2', sa.Column('t_remarks', sa.String(length=100), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('tbl_endorsement_t2', 't_remarks')
+    op.drop_column('tbl_endorsement_t1', 't_remarks')
+    # ### end Alembic commands ###
