+++ app/views/validatorSchema/__init__.py
@@ -1,0 +1,1 @@
+from app.views.validatorSchema.EndorsementFormSchema import EndorsementFormSchema
